+++ Editor/TemplInputAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Willykc.Templ.Editor
+{
+    [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
+    public class TemplInputAttribute : Attribute
+    {
+    }
+}