+++ Tests/Editor/TemplCoreNoSettingsTests.cs
@@ -38,7 +38,7 @@ namespace Willykc.Templ.Editor.Tests
         private SettingsProviderMock settingsProviderMock;
         private Type[] typeCache;
         private TemplSettings settings;
-        private AssetChanges changes;
+        private AssetsPaths changes;
         private EntryMock firstEntryMock;
 
         [OneTimeSetUp]
@@ -52,7 +52,7 @@ namespace Willykc.Templ.Editor.Tests
         [SetUp]
         public void BeforeEach()
         {
-            changes = new AssetChanges(new string[0], new string[0], new string[0], new string[0]);
+            changes = new AssetsPaths(new string[0], new string[0], new string[0], new string[0]);
 
             subject = new TemplCore(
                 assetDatabaseMock = new AssetDatabaseMock(),
@@ -102,7 +102,7 @@ namespace Willykc.Templ.Editor.Tests
         public void GivenNoSettings_WhenAssetDeleted_ThenShouldNotFlagEntries()
         {
             // Setup
-            firstEntryMock.willDelete = true;
+            firstEntryMock.inputChanged = true;
 
             // Act
             subject.OnWillDeleteAsset(string.Empty);