+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Declare custom entry input field using TemplInputAttribute
+- Declare custom entry metadata using TemplEntryInfoAttribute
 - Declare custom functions using TemplFunctionAttribute on static classes
+- Only entry types with a single input field which is subclass of Unity.Object are allowed
 
 ## [0.1.0-preview] - 2022-07-16